Separate line, extrema, and find tests
